cgroup: introduce per cgroup timer slack.
Cgroups can now have different timer slack values (cgroup.timer_slack_ns). The
timer slack is inherited down to the descendant cgroups, that can override the
inherited value for their own subtree of descendants, if necessary. A process
that hasn't changed its timer slack value through the appropriate prctl, will
the cgroup provided one which can be either shorter or longer than the default
50 us timer slack previously used in Linux. The 50 us timer slack will still
remain as the default timer slack if the cgroup values are left untouched.

Example inheritance in a cg-hierarchy that has a new timer slack set as R on
the root cgroup, in addition to N and O set in the corresponding descendants:

       {R,s}
      /     \
   {N,r}   {_,R}
    /        \
  {_,N}     {O,r}
    |         |
  {_,N}     {_,O}

where {X,y} denotes the timer slack (X) and the inherited slack (y). The
effective timer slack is in upper case, e.g. {_,Y} means the default inherited
timer slack (Y) is used. Underscore (_) denotes a default timer slack, in which
case the inherited timer slack is used.

Signed-off-by: Roger Blomgren <roger.blomgren@iki.fi>