hrtimer: Allow running (soft) timers opportunistically after soft timeout.
The softirq_expires_next is the least hard timeout value (timeout + slack) for
the base, but there can be timers where timeout (sans slack) < now. As the
timers are now sorted in softexpires order, we get the next timer cheaply and
might as well run it if it's available, possibly avoiding a wakeup from idle
later.

Signed-off-by: Roger Blomgren <roger.blomgren@iki.fi>